Share judgment_fn parser between formality-coverage and formality-mdbook
